+ Column patch
The column is now represented by the boundary of its bottom face instead of the location point anymore. Rely the geometry on column type label may not be robust. However, the column geometry may be subtracted by walls, and there may be special-shaped/T-shaped/round columns, so once again the bounding box serves the boundary for checking potential joining walls.